Changed the event handler structure.  Event::HandlerBase, Event::Handler and Event::Handler1 have replaced EventHandlerBase, EventHandler and EventHandler1Arg, repectively.  The new classes are only templatized on the class creating the handler (or rather the class containing the function being pointed to).
Also replaced LinkAdd and selfLink with configureLink and
configureSelfLink.  The only real change is that you can now set the
link's timebase as part of the configure call instead of having to do
it separately (though you can still do it separately).  Also made the
setting of link time bases by the registerClock and registerTimeBase
consistent.  If you set the registered clock or timebase as the
default, it will set all Links to that timebase as long as their
timebase has not already been set.

Note that clock handlers still use the old templatized classes, but
this will be changed to mirror the event handling in the near future.