some general cleanup to coincide with moving the the gem5 hg repository
The gem5 simulate function now return an pointer to an event for all cases. 
To differentiate between a gem5 app exit and simulate reaching a given tick 
simulate now sets the exit code to a value "256" that cannot be returned by
an application call to exit().

Move power related code to power.cpp. There was not attempt to make the 
moved code work.