Restore the portals4 test to survive the SST namespace enhancement.
Re-enable the test to run in nightly.   It was running just in  Open MPI 1.4.4 tests, simply to reduce the total amount time the test consumed each night.  Open MPI 1.4.4 tests went away.  It is now to run in Open MPI 1.6 test on Ubuntu and CentOS.  (MacOS is excluded for other reasons.)

Inserts a Begin and End flag lines around dotests() in Bamboo, to make it easier to extract dotests() from the current Bamboo for testing changes to the bamboo test driver.