Changed some of the interfaces for Merlin.  Speeds are now specified in B/s or b/s and sizes are specified in B or b.
It is no longer necessary to specify num_vns for routers or LinkControls.  You
now specify number of virtual networks (num_vns) for the LinkControls and this
information is propagated to the routers, which then make the correct number
of VCs to route deadlock free with the requested VNs.

This will break the build for anything using LinkControl and will break the
input decks for anything using any part of merlin.