Added a rank-level check for interthread dependencies.
Previously, each thread only looked at its own dependencies to determin
whether or not to instance an EmptyThreadSync object or an active one.
That lead to deadlock in the case where some, but not all, threads had no
dependencies with other threads on the rank.  The rank-level check should
fix this.