Added find_*_array functions to Params.
The new find_*_array functions in Params will get arrays of items. the arrays
are stored as strings where the first character must be [ and the last character
must be ].  The items are then comma separated inside the brackets.  This is the
format that you naturally get by passing a python list into the addParam call
in the python interface to components.  Will need to add more checking to the
functions in the future.