Added enhancement to guarantee same event delivery order for serial and parallel.
Added an extra sort criteria to Activity when SST_ENFORCE_NAME_ORDERING is set.
This is currently set in activity.h, but could be moved to sst_config.h and
be made controllable by the configure scripts.  When set, events will add in
the link ID of the last link to send it into the sort criteria.  This addition
adds no additional memory requirement (Link was already storing this data, the
location is just moved when the flag is on).  The new sorting sorts on
delivery time, then priority, then link ID and if using the priority queue
version of TimeVortex, will then sort on queue_order.  The implication of
this change is that lower numbered links will get priority delivery into
components.