Fix to simultion end timing.
Simluation::currentSimCycle is now set to Simulation::endSimCycle
before finish() is called on elements.  This makes it so that any
calls to get the core simulation time during finish will return the
actual end time instead of the time the TimeVortex advanced to before
it realized the simulation was done.