New ELI now works for partitioners.
Added Partitioners to the new ELI and moved all the built-in
partitioners to use the new ELI instead of it's own database.
Built-in partitioners now live in the sst library in the new
ELIL database.  We could also add the core modules here in
the future.  To make this backward compatible, changed the
Config class to add sst. to any partitioner name that does
not already contain a library name.