Added python modules to new ELI and other ELI tweaks.
Added SSTElementPythonModule which simplifies creation of a python
module in an element library.

Changed the way custom create functions work for Components and
SubComponents.  Now, you call a different macro if you want to
do a custom create function and implement the create function
in your class.  This gets rid of the template metaprogramming
that was used to look for a custom create funciton.

Also cleaned up a bunch of debug printouts and streamlined the
ELI printing.