Changed how subcomponents get named.
Subcomponents now get a hierarchical name.  The first part of the
name is the Component name, then each level of subcomponent is
names by slot name and separated by a colon.  Also, if there is
more than one subcomponent in a slot, then they are differentiated
by [index] appended to the slot name.