Added SubComponentSlots fully to new ELI.
Elements that use new ELI must now define the DOCUMENT_SUBCOMPONENT_SLOT
call.  When loading a named slot, the core will check to see if it's
documented and will issue a warning if it's not.  It is no longer
necessary to have the libname_eli symbol in the library if you use
the new ELI.