Changed deprecated warnings for old ELI.
Changed the deprecated for old ELI to make clear that the old ELI
is still the preferred API until 7.1, but the warning starts the
promised one major release countdown to removal.

Also added the ability for elements to turn off the warnings by
using the correct #define to turn it off.  This can be used to
turn off the warnings for our supported components, but have them
show up on external components.