Convert old ELI structures to new ELI structures.
On library load, the factory will now create the new ELI data structures
and them to the ELI Database, instead of using the old data structures.
The only thing that will not translate are generators, so they are still
stored as they have been.  Also, events are ignored, since they are not
supported in the new ELI and are not generally used.  Factory::RequireEvent()
now just loads the library and does not look for the actual event.