Added an iterative complete() stage following the run() stage.
As part of this, added calls sendUntimedData and recvUntimed data.
There will eventually replace the send/recvInitData calls to keep
the names more generic.  As part of this, changed all variables with
init in the name to use untimed instead (except for those which were
referring specifically to the init() stage.