Added UnitAlgebra support in the python input files. Added test python file to test this functionality.
Along the way, made the following fixes/changes:

- Added macro calls to get rid of deprecation warnings for python 3.8
when filling in the object data structure.  The issues is with
tp_print, that was moved to the end of the structure and deprecated.
Some compilers will warn if it's not there and some will warn that
it's deprecated if it is.  It's now added where it needs to be, but
deprecation warnings are disabled for that case.

- Fixed issue in decimal_fixedpoint where signs was computed correctly
(or at all) in multiply and divide.

- Made UnitAlgebra::init function public.