First attempt at metis partitioner
This is a rough draft of a metis partitioner for SST. The partitioner
takes the `PartitionGraph` and partitions it using Metis'
`METIS_PartGraphKway`. This is a serial partitioner, but we have found
in practice it's able to partition graphs with up to millions of edges
in a reasonable amount of time.

The partition is currently unweighted as it assigns a weight of one
per component, and a weight of 1 to each link. Simple statistics such
as edge cut fraction and imbalance are outputted.

The `CSRMat` class is taken from a different project and thus may be
slightly overdesigned. The main idea is to take a set of weighted
vertices and a set of weighted edges. These are then formatted in a
manner that is parsable by Metis.

Todos
 - Need to update header guard for `util.h`
 - Need to remove dependence on 64-bit version of Metis