Add `typename` to some uses of std::conditional.
This is needed when at least one of the two types is a dependent type. For
example, while `std::conditional<true, int, float>::type` is fine, when `T` is a
template parameter, `typename std::conditional<true, int, T>::type` is
necessary.

These replacements are entirely inside of the ELI macros that use them, and
allow us to use template classes in our element hierarchies.