Use Meyers singletons (#1222)
* Use Meyers singletons (static variables local to functions which are
initialized on their first call and return a reference to it; avoids
static initialization order fiasco) instead of pointers allocated on
demand.

Use std::shared_ptr for LoadedLibraries loaded libraries so that
libraries and aliases can share the same loader pointers without
memory leaks.

Use C++17 "inline" with static const members so that they can be
initialized in-class and won't have multiple definition errors if
defined in a header.