Add override for virtual functions which are overriden (as warned by -Wsuggest-override).
For default constructors with all bases and class members default-initialized
and an empty constructor body, and for destructors with an empty body, move the
definiton from the .cc file to the .h file with a = default definition.

Add "explicit" to constructors with a single parameter which should not perform
implicit conversions. I built sst-core and sst-elements and got no errors, but
if some of the constructors need to perform implicit conversions, then explicit
can be removed.

Add = delete declarations for copy constructors and copy assignment operators
for classes which do not need them (or which had marked them as private). Not
all classes which can have deleted copy constructors and assignment operators
were modified by this, but many were.

Add virtual destructors to classes with virtual functions and no declared
destructor.

Add __attribute__((format(...))) to printf-style functions which take format
strings. For non-static member functions, the "this" pointer is implicitly the
first parameter, so the 2nd and 3rd arguments of the format() attribute are
increased by one.

Add [[noreturn]] to non-virtual functions which do not return, such as calling
Output::fatal(). (If the function is virtual, then [[noreturn]] has little
benefit unless the function is marked as "final", which is dangerous to do
because SST Elements and other client code might override SST Core classes, and
without a careful analysis of which functions and classes are really "final" in
SST Core, it is too risky to add "final" to functions or classes, so
[[noreturn]] is left off of virtual functions.)

Use std::make_unique() for variable-length arrays instead of array[N] when N is
not constant. VLAs are non-standard, and allocate on the stack instead of the
heap, which increases the chance of stack overflows.

Replace volatile counters with std::atomic types, except in one case where the
assembly code is listed and analyzed and "volatile" is used for a timing delay.

Reformat SST_ELI_REGISTER_SUBCOMPONENT_API() macro calls to have one argument
per line.

Reformat long TimeVortex string messages to be split across lines.

Reformat sst-config print_usage() to use one call to fputs() instead of dozens
of consecutive calls to fprintf().

Rename sst-info's dprintf() to sst_dprintf() since dprintf() is a POSIX C
function which prints to an integer file descriptor, while sst_dprintf() is a
fprintf() taking a FILE pointer which prints only if verbosity is enabled
(perhaps it was named "dprintf()" to mean "Debug printf()", but dprintf() is a
name already used elsewhere).

Could have been done, but wasn't:

Default constructors and destructors defined with empty bodies in the header
could be replaced with = default which is more specific than {} and could
conceivably be more optimized. To minimize diffs, these {} default constructors
and destructors in headers were left alone. Only if they were newly added or
moved from .cc to .h files, was = default used instead of {}.

"virtual" could be removed wherever "override" was used. "override" implies
"virtual", but it was felt that leaving "virtual", which is at the begining,
would make it more readable. If "virtual" existed before, it was not removed
simply because "override" was added, but if neither "virtual" nor "override"
existed and "override" applied because the function was virtual in a base
class, then only "override" was added.