Add override for virtual functions which are overriden (as warned by -Wsuggest-override).
"virtual" could be removed wherever "override" was used. "override" implies
"virtual", but it was felt that leaving "virtual", which is at the begining,
would make it more readable. If "virtual" existed before, it was not removed
simply because "override" was added, but if neither "virtual" nor "override"
existed and "override" applied because the function was virtual in a base
class, then only "override" was added.