Add __attribute__((format(...))) to printf-style functions which take format (#1256)
strings. For non-static member functions, the "this" pointer is implicitly the
first parameter, so the 2nd and 3rd arguments of the format() attribute are
increased by one.

Rename sst-info's dprintf() to sst_dprintf() since dprintf() is a POSIX C
function which prints to an integer file descriptor, while sst_dprintf() is a
fprintf() taking a FILE pointer which prints only if verbosity is enabled
(perhaps it was named "dprintf()" to mean "Debug printf()", but dprintf() is a
name already used elsewhere).