Inline trivial default constructors and destructors (#1260)
* For default constructors with all bases and class members default-initialized
and an empty constructor body, and for destructors with an empty body, move the
definiton from the .cc file to the .h file with a = default definition.

Default constructors and destructors defined with empty bodies in the header
could be replaced with = default which is more specific than {} and could
conceivably be more optimized. To minimize diffs, these {} default constructors
and destructors in headers were left alone. Only if they were newly added or
moved from .cc to .h files, was = default used instead of {}. This makes the
new constructors/destructors potentially inline-able.

* clang-format