Add new is_same_type_template_v<> template and simplify some template type (#1303)
is_same_type_template_v<TYPE, TEMPLATE> returns true if TYPE is a variation
of TEMPLATE with arguments.

The name is *_type_template_* to emphasize that the arguments are a type and a
template, respectively, in that order.

Use is_same_type_template_v<> to simplify tuple, adapter and insertable
container code

Merge pointer and non-pointer serialize_impl adapter implementations

Add is_insertable_v<> template to test whether a type is an insertable type,
and use it to simplify pointer and non-pointer serialize_impl implementations
for insertable containers