TimeVortex is no longer serialized during checkpoint/restart. (#1334)
On checkpoint, the TimeVortex returns a copy of its contents, which are
then sorted. Events are serialized with the Component trees they are
targetting. Actions are ignored and are recreated on restart. The Actions
that are recreated are: StopAction, Exit, Clocks, OneShots (for stats).

As a consequence, a couple things were added:
- BaseComponent now sorts the clocks that are registered through it so
  that it can ensure they exist after restart
- OneShotManager was added and can create OneShots whose trigger time
  is either absolute or relative to current time

Stat Groups now checkpoint/restart correctly

Event and Action now define both isAction() and isEvent() as final

Unrelated fix: TimeVortexPQ can now properly print its contents