Test framework: deprecate passing bytes to os_awk_print (#1445)
While it is possible for the function to operate on bytestrings in addition to
(Unicode) strings, it is unlikely that bytestrings will be used as the
captured stdout from subprocess calls should always converted to strings for
use in other functions anyway.

This deprecation will put the future param type check in line with the
function type annotation.